[JTL] Stale XP Awarding fix
Fixed an issue where a character can enter a "stale" state where they are unable to receive starship xp even if they have  0 out of whatever their current cap is.

+++ MMOCoreORB/src/server/zone/objects/ship/ai/ShipAgentTemplate.cpp
@@ -69,7 +69,7 @@ void ShipAgentTemplate::readObject(LuaObject* templateData) {
 	pilotTemplate = templateData->getStringField("pilotTemplate").trim();
 	shipType = templateData->getStringField("shipType").trim();
 
-	experience = templateData->getIntField("experience");
+	experience = (int)templateData->getFloatField("experience");
 
 	lootChance = templateData->getFloatField("lootChance");
 	lootRolls = templateData->getIntField("lootRolls");
